move helper functions from google.go
